fix(db): widen chain_id to BIGINT — a TRON testnet id does not fit INTEGER
TRON derives its chain id from the low 4 bytes of the network's genesis hash
read as UNSIGNED. Mainnet's 728126428 happens to fit a 4-byte signed integer.
Nile's 3448148188 does not: Postgres INTEGER stops at 2147483647.

Both columns the TRON poller writes were Integer — indexer_cursors.chain_id
(the per-network cursor's primary key) and payment_settlements.chain_id
(indexed, part of uq_settlement_onchain_log).

This is the failure mode the test suite is structurally unable to see. Tests
and CI run on sqlite+aiosqlite, which is dynamically typed and stores the
value without complaint, so the models, the whole suite and the CI job would
all be green while production raised NumericValueOutOfRange on the first Nile
cursor write. Verified by hand against a real Postgres: with 0020 the value
round-trips through both columns; downgraded to 0019 the same insert fails
with "integer out of range". The regression test therefore lives in the
ALEMBIC_TEST_DATABASE_URL-gated file, and asserts the stored value AND the
declared data_type so it cannot pass for the wrong reason.

The downgrade refuses while an oversized id is stored rather than narrowing.
Narrowing is not reversible for those rows — the value either errors mid-ALTER
or wraps to a negative number matching no chain, silently reassigning a
settlement to nowhere. It names the blocking rows and stops; pinned in both
directions, refusal and success-once-cleared.

+++ services/backend/app/models/indexer_models.py
@@ -8,7 +8,7 @@ the merchant's wallet). Redis remains a write-through hot cache only.
 """
 from datetime import datetime, timezone
 
-from sqlalchemy import BigInteger, Column, DateTime, Integer
+from sqlalchemy import BigInteger, Column, DateTime
 
 from app.models.db_models import Base
 
@@ -17,7 +17,14 @@ class IndexerCursor(Base):
     __tablename__ = "indexer_cursors"
 
     # One row per chain; chain_id is the natural primary key.
-    chain_id = Column(Integer, primary_key=True, autoincrement=False)
+    #
+    # BigInteger, not Integer (widened in migration 0020): chain ids are not all
+    # small. TRON's are the low 4 bytes of the network's genesis hash read as
+    # UNSIGNED, and Nile's — 3448148188 — is above the 2147483647 ceiling of a
+    # Postgres INTEGER. SQLite would have stored it regardless, so this is a
+    # column the test suite cannot defend on its own; see
+    # test_migrations_postgres.py::test_a_tron_testnet_chain_id_fits_after_0020.
+    chain_id = Column(BigInteger, primary_key=True, autoincrement=False)
     last_block = Column(BigInteger, nullable=False)
     updated_at = Column(
         DateTime(timezone=True),
